move code into src dir
