+++ src/app.js
@@ -3,10 +3,8 @@
 
 const http = require('http')
 const path = require('path')
-
 const debug = require('debug')('laser:web')
 const express = require('express')
-
 const routes = require('./routes')
 
 const app = express()
@@ -18,7 +16,7 @@ require('./socket').init(server)
 let init = false
 
 // view engine setup
-app.set('views', path.join(__dirname, 'views'))
+app.set('views', path.join(__dirname, '..', 'views'))
 app.set('view engine', 'jade')
 
 module.exports.addHandler = function (path, handler) {
@@ -33,7 +31,7 @@ module.exports.app = function () {
 
     routes.addErrorHandlers(app)
 
-    app.use(express.static(path.join(__dirname, 'public')))
+    app.use(express.static(path.join(__dirname, '..', 'public')))
 
     // catch 404 and forward to error handler
     app.use(function (_req, _res, next) {
