+++ test/testutil.js
@@ -7,7 +7,7 @@ const sinon = require('sinon')
 let init
 
 module.exports.getApp = function () {
-  const mainApp = require('../app')
+  const mainApp = require('../src/app')
   if (!init) {
     mainApp.addHandler('/mock500', function (_req, _res, next) {
       next('Unittest error')
